Moved serializers and tests to api.
